Update websocket redial logic ========================
This fixes a bug that caused endless redialling if the server went down, rather than
the internet connection, and then recovered. This was caused by continuing to use
the out-of-date websocket connection in the key listen goroutine. Now it uses a
to exit + restart the key listen goroutine when the websocket connection is
sucessfully redialled, and make sure it’s using a fresh websocket connection when
it attempts to send a message.